+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/regionals/RegionalsTeleop.java
@@ -1,7 +1,6 @@
 package org.firstinspires.ftc.teamcode.regionals;
 
 import static org.firstinspires.ftc.teamcode.meet2.Meet2Auto.shooterSpeedGap;
-import static org.firstinspires.ftc.teamcode.pedroPathing.Tuning.draw;
 import static org.firstinspires.ftc.teamcode.pedroPathing.Tuning.follower;
 
 import com.bylazar.configurables.annotations.Configurable;
@@ -11,9 +10,9 @@ import com.bylazar.telemetry.TelemetryManager;
 import com.bylazar.utils.LoopTimer;
 import com.pedropathing.geometry.Pose;
 import com.qualcomm.hardware.lynx.LynxModule;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+import com.qualcomm.robotcore.hardware.AnalogInput;
 import com.qualcomm.robotcore.hardware.ColorSensor;
 import com.qualcomm.robotcore.hardware.DistanceSensor;
 import com.qualcomm.robotcore.hardware.PwmControl;
@@ -36,7 +35,7 @@ import org.firstinspires.ftc.teamcode.Meet_ILT.Drawing;
 import org.firstinspires.ftc.teamcode.Meet_ILT.ILT_Auto;
 import org.firstinspires.ftc.teamcode.meet3.Meet3Auto;
 import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
-import org.firstinspires.ftc.teamcode.prototyping.LaunchAngleCalculator;
+import org.firstinspires.ftc.teamcode.regionals.RegionalsAuto;
 import org.firstinspires.ftc.vision.VisionPortal;
 
 import java.util.List;
@@ -44,21 +43,24 @@ import java.util.Objects;
 
 @TeleOp(name = "Regionals Teleop")
 @Configurable
-@Disabled
-public class regionalTeleop extends LinearOpMode {
+public class RegionalsTeleop extends LinearOpMode {
+
     @IgnoreConfigurable
     static TelemetryManager telemetryM;
 
     //April Tag Variables
-    private boolean isFull = false;
+
 
     //Pedropathing Variables
     private Pose currentPose;
 
     private final ElapsedTime launchTimer = new ElapsedTime(ElapsedTime.Resolution.SECONDS);
+    private final ElapsedTime intakeTimer = new ElapsedTime(ElapsedTime.Resolution.SECONDS);
+    private final ElapsedTime launchPrepTimer = new ElapsedTime(ElapsedTime.Resolution.SECONDS);
+
 
     private MotorEx fL, fR, bL, bR, launcher1, launcher2, turret, intake;
-    private ServoEx hoodServo;
+    private ServoEx hoodServo, blockerServo;
     private DistanceSensor distanceSensor1, distanceSensor2;
     private ColorSensor colorSensor;
     private String DTState="drive", intakeState="idle", turretState="idle", launcherState="idle", hoodState="idle";
@@ -69,8 +71,7 @@ public class regionalTeleop extends LinearOpMode {
 
 
     //Team Dependents
-    public static String team = Meet3Auto.team;
-    private double goalID = 20;
+    public static String team = RegionalsAuto.team;
     private Pose startPose;
     private Pose goalPose;
     private Pose aprilTagPose;
@@ -93,7 +94,7 @@ public class regionalTeleop extends LinearOpMode {
     public static double ki = 200;
     public static double kd = 0;
     private double launcherTargetVelocity;
-    public static double launcherTestSpeed = 0.65;
+    public static double launcherTestSpeed = 0.25;
     public double odoRange = 0;
     public boolean isIdle = true;
 
@@ -104,8 +105,8 @@ public class regionalTeleop extends LinearOpMode {
     public static double hoodOffset = -1.5;
 
     //Blocker Variables
-    public double closedAngle = 0;
-    public double openAngle = 180;
+    public double closedAngle = 250;
+    public double openAngle = 315;
 
     //Intake Variables
     public static double intakePickupSpeed = 1.0;
@@ -114,19 +115,20 @@ public class regionalTeleop extends LinearOpMode {
 
     //Turret Variables
     private PIDFController turretPIDF;
-    private PIDController launcherPID;
     public static double turretTolerance = 2;
-    public static double tkP = 0.0025;
+    public static double tkP = 0.0045;
     public static double tkI = 0;
-    public static double tkD = 0.00005;
-    public static double tkSCustom = 0.13;
+    public static double tkD = 0.0003;
+    public static double tkSCustom = 0.08;
     public static double errorTotal = 30;
-    private int turretTargetPos;
+    private double turretTargetPos;
     private double angleError;
     private boolean turretManualControl = false;
     private boolean turretAngleLimited = false;
     private double turretDriftOffset = 0;
     private boolean driftAdjustToggle = false;
+    private boolean loadAdjustToggle = false;
+    private double loadSpeedOffset = 0;
 
     private double distance1=0;
     private double distance2=0;
@@ -141,14 +143,25 @@ public class regionalTeleop extends LinearOpMode {
     private double voltage;
     private double voltageMultiplier;
 
-    double turret_x;
-    double turret_y;
-
     //The variable to store our instance of the vision portal.
 
     private VisionPortal visionPortal;
 
     private Drawing drawing;
+
+    //Analog Encoder Variables
+    AnalogInput absEncoder;
+    public static double zeroPositionOffset = -186;
+    double outputVoltage=0;
+    double outputAngle=0;
+    double totalRawAngle = 0;
+    double turnCounter = 0;
+    double trueTurretAngle = 0;
+    double previousAngle = 0;
+    double previousRawAngle = 0;
+
+    public static double angleVelocityMultiplier = 0.05;
+    double hoodCompensationAngle = 0;
 
     @Override
     public void runOpMode() {
@@ -196,8 +209,8 @@ public class regionalTeleop extends LinearOpMode {
             }
         }
         updateTeamDependents();
-        if (ILT_Auto.endPosition.getX() != 0 && !useRealStart) {
-            startPose = ILT_Auto.endPosition;
+        if (RegionalsAuto.endPosition.getX() != 0 && !useRealStart) {
+            startPose = RegionalsAuto.endPosition;
         } else {
             try {
                 if (endPose.getX() != 0 && !useRealStart){
@@ -205,17 +218,11 @@ public class regionalTeleop extends LinearOpMode {
                 }
             } catch (NullPointerException ignored) {
             }
-            try {
-                if (startPose.getX() == 0 && useRealStart){
-                    turret.resetEncoder();
-                }
-            } catch (NullPointerException ignored) {
-            }
         }
-        if (useRealStart) turret.resetEncoder();
         follower.setStartingPose(startPose);
         if (opModeIsActive()) {
             teleTimer.reset();
+            previousAngle = absEncoder.getVoltage() / 5 * 360;
             while (opModeIsActive()) {
                 time.start();
                 hubs.forEach(LynxModule::clearBulkCache);
@@ -236,21 +243,17 @@ public class regionalTeleop extends LinearOpMode {
     }   // end method runOpMode()
 
     public void telemetryUpdate() {
-        telemetryM.addLine("Robot Position");
+//        telemetryM.addLine("Robot Position");
         telemetryM.addData("X",currentPose.getX());
         telemetryM.addData("Y",currentPose.getY());
         telemetryM.addData("Heading",Math.toDegrees(currentPose.getHeading()));
         telemetryM.addData("Goal Distance",odoRange);
-        telemetryM.addLine("Launcher Info");
-        telemetryM.addData("Launcher Velocity",launcher.getVelocity());
-        telemetryM.addData("Launcher Target",velocityLUT.get(launcherTargetVelocity));
-        telemetryM.addData("Launcher Power",launcher.get());
-        telemetryM.addLine("Turret Info");
-        telemetryM.addData("Turret Target",turretTargetPos);
-        telemetryM.addData("Turret Pos",turret.getCurrentPosition());
-        telemetryM.addData("Turret Angle",turret.getCurrentPosition() * 360/1076.6);
-        telemetryM.addData("Turret Motor Power",turret.get());
-        //telemetryM.addData("Distance",testDistanceSensor.getDistance(DistanceUnit.INCH))
+        telemetryM.addData("Turret Angle",trueTurretAngle);
+        telemetryM.addData("Hood Angle",hoodTargetAngle);
+        telemetryM.addData("Ball in 1?", ballIn1);
+        telemetryM.addData("Ball in 2?", ballIn2);
+        telemetryM.addData("Ball in 3?", ballIn3);
+
         drawing.drawRobot(new Position(DistanceUnit.INCH, currentPose.getX(),currentPose.getY(), 0,0), new YawPitchRollAngles(AngleUnit.RADIANS,turret.getCurrentPosition(), 0,0,0));
     }
 
@@ -266,6 +269,7 @@ public class regionalTeleop extends LinearOpMode {
         //turret.resetEncoder();
         turret.setZeroPowerBehavior(Motor.ZeroPowerBehavior.BRAKE);
         turret.setRunMode(Motor.RunMode.RawPower);
+        turret.setCachingTolerance(0.01);
         launcher2.setInverted(true);
         launcher1.setVeloCoefficients(kp,ki,kd);
         launcher2.setVeloCoefficients(kp,ki,kd);
@@ -285,6 +289,9 @@ public class regionalTeleop extends LinearOpMode {
         hoodServo = new ServoEx(hardwareMap,"hoodServo",0, 300);
         hoodServo.setPwm(new PwmControl.PwmRange(500,2500));
         hoodServo.setInverted(true);
+        blockerServo = new ServoEx(hardwareMap,"blockerServo",0,300);
+        blockerServo.setPwm(new PwmControl.PwmRange(500,2500));
+        absEncoder = hardwareMap.get(AnalogInput.class,"AbsoluteEncoder");
     }
 
     public void createLUTs() {
@@ -292,12 +299,25 @@ public class regionalTeleop extends LinearOpMode {
 
         //Add values (obtained empirically)
         //Input is distance, output is shooter velocity
+//        rangeLUT.add(20, 0.43);
+//        rangeLUT.add(48, 0.44);
+//        rangeLUT.add(55, 0.45);
+//        rangeLUT.add(65.7, 0.48);
+//        rangeLUT.add(75, 0.5);
+//        rangeLUT.add(85, 0.54);
+//        rangeLUT.add(100, .57);
+//        rangeLUT.add(110, 0.6);
+//        rangeLUT.add(115, 0.62);
+//        rangeLUT.add(123, 0.69);
+//        rangeLUT.add(130, 0.69);
+//        rangeLUT.add(140, 0.76);
+//        rangeLUT.add(160, 0.81);
         if (speedOverride == 0) {
             rangeLUT.add(47,0.4);
             rangeLUT.add(55,0.41);
             rangeLUT.add(65,0.44);
             rangeLUT.add(75,0.47);
-            rangeLUT.add(85,.47);
+            rangeLUT.add(85,.485);
             rangeLUT.add(95,0.5);
             rangeLUT.add(106,0.52);
             rangeLUT.add(126,.56);
@@ -322,25 +342,30 @@ public class regionalTeleop extends LinearOpMode {
         velocityLUT.add(1.001,2500);
         velocityLUT.createLUT();
 
-        hoodLUT.add(0, 33);
-        hoodLUT.add(20, 33);
-        hoodLUT.add(47,33);
-        hoodLUT.add(55,33);
-        hoodLUT.add(65,35);
-        hoodLUT.add(75,36);
-        hoodLUT.add(85,39);
-        hoodLUT.add(95,40);
-        hoodLUT.add(106,42);
-        hoodLUT.add(126, 44.5);
-        hoodLUT.add(147, 47.5);
-        hoodLUT.add(160,49);
+        if (angleOverride == 0) {
+            hoodLUT.add(0, 33);
+            hoodLUT.add(20, 33);
+            hoodLUT.add(47,33);
+            hoodLUT.add(55,33);
+            hoodLUT.add(65,35);
+            hoodLUT.add(75,36);
+            hoodLUT.add(85,39);
+            hoodLUT.add(95,40);
+            hoodLUT.add(106,42);
+            hoodLUT.add(126, 44);
+            hoodLUT.add(147, 47);
+            hoodLUT.add(160,48.5);
+        } else {
+            hoodLUT.add(0,angleOverride);
+            hoodLUT.add(160,angleOverride);
+        }
         hoodLUT.createLUT();
-
     }
 
     public void teleOp() {
         DriverInput();
         if (sensorsEnabled) sensorTeleOp();
+        absEncoderTeleOp();
         DTTeleOp();
         turretTeleOp();
         launcherTeleOp();
@@ -348,8 +373,23 @@ public class regionalTeleop extends LinearOpMode {
         hoodTeleOp();
     }
 
+    public void absEncoderTeleOp() {
+        outputVoltage = absEncoder.getVoltage();
+        outputAngle = outputVoltage / 5 * 360;
+        if (Math.abs(outputAngle-previousAngle) < 180) {
+        } else if (outputAngle < previousAngle) turnCounter += 1;
+        else if (outputAngle > previousAngle) turnCounter -= 1;
+        totalRawAngle = outputAngle + turnCounter*360;
+        trueTurretAngle = ((totalRawAngle + zeroPositionOffset)/2);
+        previousAngle = outputAngle;
+        previousRawAngle = totalRawAngle;
+    }
+
+//    public double calculateHoodOffset(double velocityError) {
+//        return Math.abs(velocityError) * angleVelocityMultiplier;
+//    }
+
     public void hoodTeleOp() {
-        hoodTargetAngle = LaunchAngleCalculator.calcBestAngle(launcher.getVelocity(),odoRange, telemetryM);
         switch (hoodState) {
             case "idle":
                 hoodServo.set(hoodToServoAngle(hoodTargetAngle));
@@ -363,15 +403,10 @@ public class regionalTeleop extends LinearOpMode {
                 hoodServo.set(hoodToServoAngle(hoodTargetAngle));
                 break;
             case "adjusting":
-                hoodServo.set(hoodToServoAngle(angleLUT(odoRange)));
+                hoodTargetAngle = hoodToServoAngle(angleLUT(odoRange)) + hoodCompensationAngle;
+                hoodServo.set(hoodTargetAngle);
                 break;
-            case "firing":
-                hoodServo.set(hoodToServoAngle(hoodTargetAngle));
-                break;
-
         }
-        telemetryM.addData("Hood Angle",hoodTargetAngle);
-        //telemetryM.addData("Servo Angle",hoodServo.get());
     }
 
     public double servoToHoodAngle(double servoInput) {
@@ -387,63 +422,49 @@ public class regionalTeleop extends LinearOpMode {
         else return 50;
     }
     public void DriverInput() {
-
-//        if (gamepad1.dpad_up) {
-//            hoodState = "up";
-//        } else if (gamepad1.dpad_down) {
-//            hoodState = "down";
-//        } else if (gamepad1.dpad_left) {
-//            hoodState = "idle";
-//        }
-
-        //DT Section
+                //DT Section
         strafeInput = gamepad1.left_stick_x;
         driveInput = -gamepad1.left_stick_y;
         turnInput = gamepad1.right_stick_x;
         if (gamepad1.left_bumper) {
             drive.setMaxSpeed(0.7);
             turnInput *= (0.5/0.7);
+        } else if(gamepad1.left_trigger > 0.4) {
+            drive.setMaxSpeed(0.4);
+            turnInput *= 0.8;
         } else {
             if (Math.abs(driveInput) > 0.7 && Math.abs(strafeInput) < 0.3) strafeInput = 0;
             drive.setMaxSpeed(1);
         }
 
-        if (gamepad1.right_trigger > 0.4 && gamepad1.left_trigger > 0.4) {
+        if (gamepad1.left_trigger > 0.4 && gamepad1.right_trigger > 0.4) {
             follower.setPose(poseResetPose);
         }
-        if (gamepad2.right_trigger > 0.4 && gamepad2.left_trigger > 0.4) {
-            follower.setPose(new Pose(139.5 - poseResetPose.getX(),poseResetPose.getY(),poseResetPose.getHeading()));
-        }
-        if (gamepad2.right_stick_button && gamepad2.left_bumper) turret.resetEncoder();
 
-        if (gamepad1.left_stick_button) turretTargetPos = 0;
+        if (gamepad2.right_stick_button && gamepad2.left_bumper) turnCounter = 0;
+
+        if (gamepad1.left_stick_button) { //Full subsystem reset for fixing stuff
+            launcherState = "off";
+            intakeState = "idle";
+            turretState = "idle";
+        }
 
         //Intake Section
-        if (gamepad2.a || gamepad1.a) {
-            intakeState = "intaking";
-            launcherState = "rejecting";
-        } else if (gamepad2.y || gamepad1.dpad_down) {
+        if (gamepad1.a) {
+            intakeState = "beginIntaking";
+        } else if (gamepad1.dpad_down) {
             intakeState = "rejecting";
-            launcherState = "rejecting";
-        } else if (gamepad2.b || gamepad1.b) {
-            intakeState = "idle";
-            if (Objects.equals(launcherState,"rejecting")) {
-                launcherState = "idle";
-            }
-        } else if (gamepad1.right_stick_button) {
-            intakeState = "firing";
-        } else if (gamepad1.x || gamepad2.x) {
+        } else if (gamepad1.x) {
             intakeState = "idle";
             turretState = "idle";
             launcherState = "idle";
-            hoodState = "adjusting";
-            //hoodState = "idle";
         }
-        if (gamepad1.y) {
+        if (gamepad1.b) {
             intakeState = "idle";
             turretState = "tracking";
-            hoodState = "adjusting";
-//            launcherState = "preparing";
+        }
+        if (gamepad1.y) {
+            intakeState = "shortReject";
         }
 
         //Launcher Section
@@ -455,7 +476,7 @@ public class regionalTeleop extends LinearOpMode {
             ballIn2 = false;
             ballIn3 = false;
         }
-        if (gamepad1.right_bumper || gamepad2.right_bumper) {
+        if (gamepad1.right_bumper) {
             launcherState = "beginLaunchSequence";
             turretState = "aiming";
             intakeState = "idle";
@@ -465,58 +486,58 @@ public class regionalTeleop extends LinearOpMode {
         //Turret Section
         if (Math.abs(gamepad2.left_stick_x) > 0.1 || gamepad2.left_bumper) {
             turretManualControl = true;
-            if (turret.getCurrentPosition() > turretAngleToTicks(-135) && gamepad2.left_stick_x < 0) {
+            if (trueTurretAngle > -143 && gamepad2.left_stick_x < 0) {
                 turret.set(gamepad2.left_stick_x * .25);
-            } else if (turret.getCurrentPosition() < turretAngleToTicks(135) && gamepad2.left_stick_x > 0) {
-                turret.set(gamepad2.left_stick_x * .25);                                                                 //////    '                                                                                                                                     iiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiki88888888888888888888888888888888888888888888888888888888888888888888888///////////////////////////');
+            } else if (trueTurretAngle < 180 && gamepad2.left_stick_x > 0) {
+                turret.set(gamepad2.left_stick_x * .25);
             } else {
                 turret.set(0);
             }
         } else turretManualControl = false;
-        /*if (gamepad2.dpad_left) { // test
-            turretTargetPos = turretAngleToTicks(45);
-        } else if (gamepad2.dpad_right) {
-            turretTargetPos = turretAngleToTicks(-45);
-        } else if (gamepad2.left_stick_button) {
-            turretTargetPos = turretAngleToTicks(0);
-        }*/
-        if (gamepad2.dpad_left && driftAdjustToggle) { // in case turret drifts, player 2 clicks left or right for autoaim adjustment!!
+
+        if ((gamepad2.dpad_left || gamepad1.dpad_left) && driftAdjustToggle) { // in case turret drifts, player 2 clicks left or right for autoaim adjustment!!
             driftAdjustToggle = false;
-            turretDriftOffset -= 1;
-        } else if (gamepad2.dpad_right && driftAdjustToggle) {
+            turretDriftOffset -= .5;
+        } else if ((gamepad2.dpad_right || gamepad1.dpad_right) && driftAdjustToggle) {
             driftAdjustToggle = false;
-            turretDriftOffset += 1;
-        } else {
+            turretDriftOffset += .5;
+        } else if (!gamepad2.dpad_left && !gamepad2.dpad_right && gamepad1.dpad_left && gamepad1.dpad_right) {
             driftAdjustToggle = true;
+        }
+
+        if (gamepad2.a && loadAdjustToggle) {
+            loadSpeedOffset -= 0.02;
+            loadAdjustToggle = false;
+        } else if (gamepad2.b && loadAdjustToggle) {
+            loadSpeedOffset += 0.02;
+            loadAdjustToggle = false;
+        } else if (!gamepad2.a && !gamepad2.b) {
+            loadAdjustToggle = true;
         }
 
         driveAngleDegrees = Math.toDegrees(currentPose.getHeading());
     }
 
     public void sensorTeleOp() {
-        if(cyclesSinceUpdate == 3) {
-            distance1 = distanceSensor1.getDistance(DistanceUnit.INCH);
-            prevBallIn1 = ballIn1;
-            ballIn1 = distance1 < 4;
+        if (Objects.equals(launcherState,"idle") || Objects.equals(launcherState, "off")) {
+            if (cyclesSinceUpdate == 9) {
+                cyclesSinceUpdate = 0;
+                distance1 = distanceSensor1.getDistance(DistanceUnit.INCH);
+                prevBallIn1 = ballIn1;
+                ballIn1 = distance1 < 4;
+            }
+            if (cyclesSinceUpdate == 6) {
+                colorAlpha = colorSensor.alpha();
+                prevBallIn2 = ballIn2;
+                ballIn2 = colorAlpha > 70;
+            }
+            if (cyclesSinceUpdate == 3) {
+                distance2 = distanceSensor2.getDistance(DistanceUnit.INCH);
+                prevBallIn3 = ballIn3;
+                ballIn3 = distance2 < 4;
+            }
+            cyclesSinceUpdate++;
         }
-        if(cyclesSinceUpdate == 6) {
-            colorAlpha = colorSensor.alpha();
-            prevBallIn2 = ballIn2;
-            ballIn2 = colorAlpha > 70;
-        }
-        if(cyclesSinceUpdate == 9) {
-            distance2 = distanceSensor2.getDistance(DistanceUnit.INCH);
-            cyclesSinceUpdate = 0;
-            prevBallIn3 = ballIn3;
-            ballIn3 = distance2 < 4;
-        }
-        cyclesSinceUpdate++;
-
-        telemetryM.addData("Distance1", distance1);
-        telemetryM.addData("Distance2", distance2);
-        telemetryM.addData("Ball in 1?", ballIn1);
-        telemetryM.addData("Ball in 2?", ballIn2);
-        telemetryM.addData("Ball in 3?", ballIn3);
     }
 
 
@@ -527,10 +548,10 @@ public class regionalTeleop extends LinearOpMode {
                 DTisReady = true;
                 break;
             case "drive":
-                if (Math.abs(driveInput) < .1) {
+                if (Math.abs(driveInput) < .05) {
                     driveInput = 0;
                 }
-                if (Math.abs(strafeInput) < .1) {
+                if (Math.abs(strafeInput) < .05) {
                     strafeInput = 0;
                 }
                 drive.driveRobotCentric(strafeInput,driveInput,turnInput);
@@ -542,23 +563,41 @@ public class regionalTeleop extends LinearOpMode {
             case "idle":
                 intakeisReady = true;
                 intake.stopMotor();
+                blockerServo.set(openAngle);
                 break;
             case "firing":
 //                if (Math.abs(intake.getVelocity()) < 300) intake.set(1.0);
 //                else intake.set(transferLoadSpeed);
                 intake.set(transferLoadSpeed);
+                blockerServo.set(openAngle);
                 break;
+            case "beginIntaking":
+                intakeTimer.reset();
+                blockerServo.set(closedAngle);
+                intakeState = "intaking";
             case "intaking":
-                intake.set(intakePickupSpeed);
-                if (ballIn1 && ballIn2 && ballIn3 && prevBallIn1 && prevBallIn2 && prevBallIn3) {
-                    intakeState = "idle";
-                    //launcherState = "preparing";
-                    turretState = "tracking";
+                if ((!ballIn1 && !ballIn2 && !ballIn3 && sensorsEnabled) || intakeTimer.seconds() > 0.3) {
+                    intake.set(intakePickupSpeed);
+                    blockerServo.set(closedAngle);
+                    if (ballIn1 && ballIn2 && ballIn3 && prevBallIn1 && prevBallIn2 && prevBallIn3) {
+                        intakeState = "idle";
+                        turretState = "tracking";
+                    }
                 }
                 break;
             case "rejecting":
                 intake.set(intakeRejectSpeed);
+                blockerServo.set(openAngle);
                 break;
+            case "shortReject":
+                intake.set(-1);
+                intakeTimer.reset();
+                intakeState = "shortRejecting";
+                break;
+            case "shortRejecting":
+                if (intakeTimer.seconds() > 0.08) {
+                    intake.set(1);
+                }
         }
     }
 
@@ -597,11 +636,15 @@ public class regionalTeleop extends LinearOpMode {
         double goalX = goalPose.getX();
         double goalY = goalPose.getY();
         switch (launcherState){
-            case "idle":
+            case "off":
                 launcherisReady = true;
                 isIdle = true;
                 launcher.stopMotor();
                 odoRange = Math.hypot(goalX-botX,goalY-botY);
+            case "idle":
+                launcherSpinUp();
+                launcherTargetVelocity *= 0.8;
+                launcher.set(launcherTargetVelocity);
                 break;
             case "testSpeed":
                 launcherTargetVelocity = launcherTestSpeed;
@@ -615,14 +658,12 @@ public class regionalTeleop extends LinearOpMode {
                 intakeState = "idle";
                 launcherState = "aiming";
                 turretState = "aiming";
+                hoodState = "adjusting";
                 break;
             case "aiming":
                 launcherSpinUp();
                 launcher.set(launcherTargetVelocity);
                 intakeState = "idle";
-                if (Math.abs(launcher.getVelocity()) > Math.abs(velocityLUT.get(launcherTargetVelocity)) - 100) {
-                    hoodState = "firing";
-                }
                 if (Math.abs(launcher.getVelocity()) > Math.abs(velocityLUT.get(launcherTargetVelocity)) - shooterSpeedGap) {
                     launcherState = "acc_ready";
                 }
@@ -630,7 +671,7 @@ public class regionalTeleop extends LinearOpMode {
             case "acc_ready":
                 launcherSpinUp();
                 launcher.set(launcherTargetVelocity);
-                if (Objects.equals(turretState, "aimed") && follower.getVelocity().getMagnitude() < 6 && Math.abs(follower.getAngularVelocity()) < 0.5) {
+                if (Objects.equals(turretState, "aimed") && follower.getVelocity().getMagnitude() < 4 && Math.abs(follower.getAngularVelocity()) < 0.5) {
                     intakeState = "firing";
                     launcherState = "firing";
                     break;
@@ -641,14 +682,22 @@ public class regionalTeleop extends LinearOpMode {
                 if (launcher.getVelocity() > velocityLUT.get(launcherTargetVelocity)) {
                     launcher.stopMotor();
                 } else launcher.set(1);
-                if (Objects.equals(turretState,"aiming") || (follower.getVelocity().getMagnitude() > 6 && Math.abs(follower.getAngularVelocity()) < 0.5)) {
+                if (Objects.equals(turretState,"aiming") ||
+                        (follower.getVelocity().getMagnitude() > 4 && Math.abs(follower.getAngularVelocity()) < 0.5)) {
                     intakeState = "idle";
                     launcherState = "acc_ready";
                 }
                 break;
+            case "prepare":
+                launchPrepTimer.reset();
+                launcherState = "preparing";
+                break;
             case "preparing":
-                launcherSpinUp();
-                launcher.set(launcherTargetVelocity);
+                if (launchPrepTimer.seconds() > 1) {
+                    launcherSpinUp();
+                    launcher.set(launcherTargetVelocity);
+                }
+                else launcher.stopMotor();
                 break;
             case "rejecting":
                 launcherTargetVelocity = -.3;
@@ -667,9 +716,11 @@ public class regionalTeleop extends LinearOpMode {
             return rangeLUT.get(159);
         } else {
             if (70 < input && input < 90) transferLoadSpeed = 0.85 * voltageMultiplier;
-            else if (90 < input && input < 110) transferLoadSpeed = 0.7 * voltageMultiplier;
-            else if (input > 110) transferLoadSpeed = 0.55 * voltageMultiplier;
-            else transferLoadSpeed = 1;
+            else if (90 < input && input < 110) transferLoadSpeed = 0.75 * voltageMultiplier;
+            else if (input > 110) transferLoadSpeed = 0.65 * voltageMultiplier;
+            else transferLoadSpeed = .93;
+            if (transferLoadSpeed < 0.52) transferLoadSpeed = 0.52;
+            transferLoadSpeed += loadSpeedOffset;
             return rangeLUT.get(input);
         }
     }
@@ -679,13 +730,13 @@ public class regionalTeleop extends LinearOpMode {
         double botY = currentPose.getY();
         double goalX = goalPose.getX();
         double goalY = goalPose.getY();
-        odoRange = Math.hypot(goalX-turret_x,goalY-turret_y);
+        odoRange = Math.hypot(goalX-botX,goalY-botY);
         launcherTargetVelocity = calculateRangeLUT(odoRange);
     }
 
     public void turretTargetProcedure() {
         turretTrackProcedure();
-        if (!turretAngleLimited) angleError = turretTicksToAngle(turretTargetPos-turret.getCurrentPosition());
+        if (!turretAngleLimited) angleError = turretTargetPos- trueTurretAngle;
         else angleError = 100;
     }
 
@@ -693,42 +744,25 @@ public class regionalTeleop extends LinearOpMode {
         double turretAngle;
         turretAngle = calculateTurretAngle(currentPose.getX(), currentPose.getY(), Math.toDegrees((currentPose.getHeading())));
         turretAngle = turretAngleLimiter(turretAngle);
-        //if (follower.getVelocity().getMagnitude() < 2 && Math.abs(follower.getAngularVelocity()) < 0.2)
-        turretTargetPos = turretAngleToTicks(turretAngle);
+        turretTargetPos = turretAngle;
     }
 
     public double calculateTurretAngle(double botX, double botY, double botHeading) {
         double goalX = goalPose.getX();
         double goalY = goalPose.getY();
-        //double turret_offset_y = .430688;
-        double offsetMagnitude = -1;
-        turret_x = botX + offsetMagnitude * Math.cos(Math.toRadians(botHeading));
-        turret_y = botY + offsetMagnitude * Math.sin(Math.toRadians(botHeading));
         double targetAngle = Math.toDegrees(Math.atan2(goalY-botY,goalX-botX));
         targetAngle -= botHeading + 180 + turretDriftOffset;
-        telemetryM.addData("Target Angle",targetAngle);
         return targetAngle;
     }
 
     public void turretControlLoop() {
-        double output = turretPIDF.calculate(turret.getCurrentPosition(),turretTargetPos);
+        double output = turretPIDF.calculate(trueTurretAngle,turretTargetPos);
         double kSFriction;
-        if (Math.abs(turretPIDF.getPositionError()) <= turretTolerance) {
-            turret.stopMotor();
-        }
-        else {
-            kSFriction = tkSCustom * (Math.abs(turretPIDF.getPositionError()) / turretPIDF.getPositionError());
-            output += kSFriction;
-            turret.set(output);
-        }
-    }
-
-    public int turretAngleToTicks(double angle) {
-        return (int) (angle * 1076.6 / 360); //978.7 with 435
-    }
-
-    public int turretTicksToAngle(double ticks) {
-        return (int) (ticks * 360 / 1076.6); //978.7 with 435
+        kSFriction = tkSCustom * (Math.abs(turretPIDF.getPositionError()) / turretPIDF.getPositionError());
+        if (Math.abs(turretPIDF.getPositionError()) > 1) output += kSFriction;
+        else if (Math.abs(turretPIDF.getPositionError()) < 1) output = 0;
+        if (Math.abs(output) > 0.7) output = output/Math.abs(output) * 0.7;
+        turret.set(output);
     }
 
     public double turretAngleLimiter(double angleAttempt) {
@@ -738,15 +772,15 @@ public class regionalTeleop extends LinearOpMode {
         } else if (realAngle < -180) {
             realAngle += 360;
         }
-        if ((realAngle > 175 && turret.getCurrentPosition() < 0) || (realAngle < -160 && turret.getCurrentPosition() > 0)) {
+        if ((realAngle > 170 && trueTurretAngle < 0) || (realAngle < -170 && trueTurretAngle > 0)) {
             turretAngleLimited = true;
-            return turretTicksToAngle(turretTargetPos);
+            return turretTargetPos;
         } else turretAngleLimited = false;
-        if (realAngle > 160) {
+        if (realAngle > 175) {
             turretAngleLimited = true;
-            realAngle = 160;
-        } else if (realAngle < -135) {
-            realAngle = -135;
+            realAngle = 175;
+        } else if (realAngle < -145) {
+            realAngle = -145;
             turretAngleLimited = true;
         }
         return realAngle;
@@ -754,13 +788,11 @@ public class regionalTeleop extends LinearOpMode {
 
     public void updateTeamDependents() {
         if (Objects.equals(team,"blue")){
-            goalID = 20;
             startPose = new Pose(30,136,Math.toRadians(270));
             goalPose = new Pose(0,144,Math.toRadians(135));
             poseResetPose = new Pose(114,7,Math.toRadians(90)); //need to find good one
             aprilTagPose = new Pose(15,130,0);
         } else if (Objects.equals(team,"red")) {
-            goalID = 24;
             startPose = new Pose(x(30),136,a(270));
             goalPose = new Pose(144,144,Math.toRadians(45));
             poseResetPose = new Pose(30,7,Math.toRadians(90)); //need to find good one
